Reduce buildtime (#27)
* Disable unused OpenCV modules and update to 4.4.0

Disabled several OpenCV modules in the portfile to reduce build size and dependencies. Updated the OpenCV version from 4.3.0 to 4.4.0 in vcpkg.json and updated the corresponding SHA512 hash. Also removed the 'released' event from the GitHub Actions workflow trigger.

* Downgrade opencv4 port to version 4.3.0

Reverts the opencv4 port version from 4.4.0 to 4.3.0 in vcpkg.json and updates the corresponding SHA512 hash in portfile.cmake to match the new version.

* Update portfile.cmake

* Update portfile.cmake

* Update my-testing.yaml

* Update my-testing.yaml