Flathub readme (#108)
* Add README for unofficial Flatpak manifest

Introduces a README in the unsupported/flatpak directory with instructions for building the plugin from source using Flatpak, a call for community maintainers, and details on prerequisites and installation steps.

* Update Linux installation instructions in README

Added Flatpak build instructions and updated Arch Linux installation link to point to the respective README files for better clarity.